requirements.txt
kiteconnect==4.1.0
python-dotenv==1.0.0
pandas==2.0.0
schedule==1.2.0
selenium==4.9.0
webdriver-manager==3.8.6
pytz==2023.3
requests==2.28.2
python-telegram-bot==13.15